Update CLI to pass map name and num of robots
